Split Perpignan into PerpignanBoard, PerpignanGame
The board object is more passive, it doesn't have the concept of a deck
and doesn't manage whose turn it is. This will become useful when
networking is implemented; a remote client will only use a
PerpignanBoard object, only the server will use PerpignanGame.

Also add type annotations everywhere, and some other refactoring.